refactor(tests): extract shared stdout/stderr write-spy helper for CLI tests
AuthCommand, ToolsCommand, CliRootArgs, InitCommand, ConfigCommand, and
Doctor each hand-rolled the same process.stdout/stderr write mock (capture
the chunk, forward any completion callback so awaited writes resolve,
return true). Hoist it to cli/fixtures/streamWriteSpy.ts (matching the
existing cliContext.ts fixture) and have all six call the shared
spyOnStreamWrite. Behavior is identical: same callback-forwarding, same
return value; ConfigCommand's variant only forwarded the callback (no
accumulator), preserved via the optional onChunk parameter.
